requirements.txt
pandas
openai>=1.0.0
PyYAML